cleaned up code and moved colorimetrie directory
