+++ src/colorimetrie/Makefile
@@ -5,7 +5,7 @@ CFLAGS  	= -g -Wall -Wextra  `pkg-config --cflags gtk+-3.0 gmodule-2.0`
 CPPFLAGS	= -MMD
 LDLIBS		= `pkg-config --libs gtk+-3.0 gmodule-2.0`
 
-SOURCE		= main.c ../image.c stack.c colorimetrie.c
+SOURCE		= main.c ../image/image.c stack.c colorimetrie.c
 OBJ			= $(SOURCE:.c=.o)
 DEP			= $(SOURCE:.c=.d)
