+++ src/colorimetrie/images/results/circle
