+++ src/colorimetrie/images/results/circle_fill
