+++ src/colorimetrie/images/results/flood_fill
