+++ src/colorimetrie/images/results/rectangle_filled
