+++ src/colorimetrie/stack.c
