--- src/image/colorimetrie/launch
@@ -1,2 +1,0 @@
-./colorimetrie images/tests/circle.png images/results/circle_test 100 | gimp
-images/results/circle_test
